chore: move icons to folder and update imports
