+++ src/components/icons/pencil-icon.js
