+++ src/components/icons/trash-icon.js
